Extracted PetTypeService from ClientService
