+++ src/main/java/org/springframework/samples/petclinic/service/ClinicService.java
@@ -57,12 +57,6 @@ public interface ClinicService {
 	void deleteOwner(Owner owner) throws DataAccessException;
 	Collection<Owner> findOwnerByLastName(String lastName) throws DataAccessException;
 
-	PetType findPetTypeById(int petTypeId);
-	Collection<PetType> findAllPetTypes() throws DataAccessException;
-	Collection<PetType> findPetTypes() throws DataAccessException;
-	void savePetType(PetType petType) throws DataAccessException;
-	void deletePetType(PetType petType) throws DataAccessException;
-	
 	Specialty findSpecialtyById(int specialtyId);
 	Collection<Specialty> findAllSpecialties() throws DataAccessException;
 	void saveSpecialty(Specialty specialty) throws DataAccessException;
