+++ src/main/java/org/springframework/samples/petclinic/service/PetTypeService.java
@@ -1,0 +1,14 @@
+package org.springframework.samples.petclinic.service;
+
+import org.springframework.dao.DataAccessException;
+import org.springframework.samples.petclinic.model.PetType;
+
+import java.util.Collection;
+
+public interface PetTypeService {
+    PetType findPetTypeById(int petTypeId);
+    Collection<PetType> findAllPetTypes() throws DataAccessException;
+    Collection<PetType> findPetTypes() throws DataAccessException;
+    void savePetType(PetType petType) throws DataAccessException;
+    void deletePetType(PetType petType) throws DataAccessException;
+}
